separate imgui contexts for RGB/OCV
this decouples host-side imgui painting from the render loop; previously
we need to re-paint to the same global context twice, as submitting the
command buffer requires the context to be available and we need two
color spaces imgui cb submit. Now painting to both RGB and OCV color
space can happen outside of the render loop as they paint to separate
contexts.